Fix for constant in do statements
In the instance of having a "do" statement with a constant defined, and a macro inside of the do statement that tries to get the value of the constant, the constant wouldn't be defined. I simply added macro-expand-all to evaluate the constant. This may or may not be the right fix for the problem.